in arithmeticoperation.*, pass doubles in parameters for consistency

+++ arithmeticoperation.cpp
@@ -1,21 +1,21 @@
 #include "arithmeticoperation.h"
 
-double ArithmeticOperation::Add(float left, float right)
+double ArithmeticOperation::Add(double left, double right)
 {
     return left + right;
 }
 
-double ArithmeticOperation::Subtract(float left, float right)
+double ArithmeticOperation::Subtract(double left, double right)
 {
     return left - right;
 }
 
-double ArithmeticOperation::Divide(float left, float right)
+double ArithmeticOperation::Divide(double left, double right)
 {
     return left / right;
 }
 
-double ArithmeticOperation::Multiply(float left, float right)
+double ArithmeticOperation::Multiply(double left, double right)
 {
     return left * right;
 }
